+++ src/NetCore/Web/Api/wwwroot/scripts/api/webgis.options.js
@@ -123,7 +123,9 @@ webgis.usability = {
 
     show_presentation_metadata_in_copyright: true,
     show_metadata_i_button_toc: true,
-    show_link_button_in_toc: true
+    show_link_button_in_toc: true,
+
+    expandBasemapsOnAddServices: false
 };
 
 webgis.colorScheme = 'default';